pyproject.toml
[project]
name = "timeiq-core-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "celery[redis]>=5.6.2",
    "fastapi>=0.115.0",
    "google>=3.0.0",
    "google-cloud-storage>=2.18.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.35.0",
    "google-auth-oauthlib>=1.2.1",
    "google-genai>=1.66.0",
    "groq>=1.1.0",
    "httpx>=0.28.1",
    "langfuse>=3.14.5",
    "langgraph>=1.0.10",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.13.1",
    "pydantic[email]>=2.12.5",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "python-dotenv>=1.2.2",
    "rapidfuzz>=3.14.3",
    "redis>=6.4.0",
    "ruff>=0.15.5",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.32.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
]
[tool.ruff]
line-length = 88
target-version = "py311"
ignore = ["B008"]

[tool.ruff.lint.flake8-bugbear]
extend-immutable-calls = ["fastapi.Depends", "fastapi.Query", "fastapi.Path", "fastapi.Body"]

[tool.ruff.lint]
select = ["E", "W", "F", "I", "B", "UP", "N", "SIM", "C4"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.15.5",
]

[tool.mypy]
python_version = "3.11"
strict = true
disallow_untyped_defs = true
no_implicit_optional = true
